% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackingFunctions.R
\name{CheckSplits}
\alias{CheckSplits}
\title{CheckSplits}
\usage{
CheckSplits(id, id_split_list)
}
\arguments{
\item{id}{The meta-cluster id to check}

\item{id_split_list}{The list of meta-clusters currently exist}
}
\value{
The number of splits for the meta-cluster
}
\description{
A function to check number of splits and return number of splits
}
