% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackSOM.R
\name{clustWithNAs}
\alias{clustWithNAs}
\title{clustWithNAs}
\usage{
clustWithNAs(values, mcs)
}
\arguments{
\item{values}{a list of SOM nodes}

\item{mcs}{a list of meta-clusters}
}
\value{
an updated list of SOM nodes
}
\description{
Function used to tease apart with the meta-clustering result for the SOM nodes.
}
