% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{ConcatenateClusteringDetails}
\alias{ConcatenateClusteringDetails}
\title{ConcatenateClusteringDetails}
\usage{
ConcatenateClusteringDetails(tracksom.result, dat, timepoint.col, timepoints)
}
\arguments{
\item{tracksom.result}{TrackSOM object as generated by \code{\link{TrackSOM}}}

\item{dat}{Dataframe containing all cells from all time points}

\item{divide.by}{Character showing column name of the dataframe that indicate
the time point of each cell}
}
\value{
Dataframe containing the meta cluster, cluster, and tracking ids for
  all the cells
}
\description{
Given a TrackSOM object and a DataFrame containing cells from all time
points, extract all the meta clusters, clusters, and tracking id and combine
them with the cell information.
}
\examples{

library(data.table)
data_files <- sapply(c(0:4), function(i) {
  system.file("extdata", paste0("synthetic_d", i, ".fcs"), package="TrackSOM")
})
use_cols <- c("x", "y", "z")
tracksom_result <- TrackSOM(inputFiles = data_files,
                            colsToUse = use_cols,
                            nClus = 10,
                            dataFileType = ".fcs"
)

data <- lapply(c(0:4), function(i) {
  fread(system.file("extdata", paste0("synthetic_d", i, ".csv"), package="TrackSOM"))
})
data <- rbindlist(data)
tracksom_res <- ConcatenateClusteringDetails(
  tracksom.result = tracksom_result,
  dat = data,
  timepoint.col = 'timepoint',
  timepoints = c("Mock", "SYN-1", "SYN-2", "SYN-3", "SYN-4")
)

}
