% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInput.R
\name{create_flowframe}
\alias{create_flowframe}
\title{Create flowframe}
\usage{
create_flowframe(input_dat, colsToUse)
}
\arguments{
\item{input_dat}{a data.table}

\item{colsToUse}{a vector of columns to isolate from input_dat}
}
\value{
flowFrame object
}
\description{
Create flowframe
}
