% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeseriesHeatmapFunctions.R
\name{DrawTimeseriesHeatmap}
\alias{DrawTimeseriesHeatmap}
\title{DrawTimeseriesHeatmap}
\usage{
DrawTimeseriesHeatmap(dat, timepoint.col, timepoints, cluster.col,
 marker.cols)
}
\arguments{
\item{dat}{Data table. The dataset}

\item{timepoints}{Vector of time-points}

\item{timepoint.col}{Column denoting the time point column}

\item{cluster.col}{Column denoting the TrackSOM meta-cluster ID}

\item{marker.cols}{Columns denoting the markers to plot}

\item{colours}{Colour scheme for colouring the nodes. Can only be
either one of the colours specified by the viridis package
\url{https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html}
or spectral, bupu, rdylbu.}

\item{plot.width}{Width of the image for network plot export}

\item{plot.height}{Height of the image for network plot export}

\item{font.size}{Size of the font}

\item{file.format}{Image export file format}

\item{min.node.size}{Smallest node size. Only used if node.size is 'auto'}

\item{max.node.size}{Largest node size. Only used if node.size is 'auto'}

\item{legend.position}{Position of the legend, as per ggplot2.}

\item{load.temp.data}{Whether to load some intermediate calculations or not.
These intermediate files contain the details of each node and edge, and
produced by the function after the very first run.}

\item{arrow.length}{Length of the arrow}

\item{mapping}{A data.table containing the annotation (cellular population)
each meta-cluster in each time-point represents. The data.table must
contain a column meta-cluster ID column named as per cluster.col,
time-point column named as per timepoint.col, and a cell population column
named as per population.col.}

\item{population.col}{Column denoting the cell population name column. Only
used if mapping is not NULL.}

\item{cluster.ordering}{A custom ordering for the meta-clusters (y-axis)}

\item{map.cluster.id}{Whether to map the meta-cluster ID to numeric. Handy if
the meta-cluster IDs get too long.}

\item{calculation.type}{The aggregation (mean/median) type for colouring the
nodes}

\item{line.width}{Width of each edge}

\item{node.size}{Whether the nodes are to be sized based on the proportion of
cells it captured ('auto') or just constant (a number denoting the node
size)}

\item{arrow.head.gap}{Gap between the node and the head of the arrow}

\item{no_merge}{Whether meta-cluster merging was allowed. This only affect
the network plot coloured by the meta-cluster origin. If merging was
permitted, any merged meta-clusters will be coloured grey. If merging was
not permitted, any split meta-clusters will be coloured the same as their
parent.}
}
\description{
Draw TrackSOM network plot where each node represents a meta-cluster,
coloured by the time-point it exists or median/mean marker values.
Meta-clusters are connected with edges, indicating a transition over time.
}
\examples{

library(data.table)
data_files <- sapply(c(0:4), function(i) {
  system.file("extdata", paste0("synthetic_d", i, ".fcs"), package="TrackSOM")
})
use_cols <- c("x", "y", "z")
tracksom_result <- TrackSOM(inputFiles = data_files,
                            colsToUse = use_cols,
                            nClus = 10,
                            dataFileType = ".fcs"
)

data <- lapply(c(0:4), function(i) {
  fread(system.file("extdata", paste0("synthetic_d", i, ".csv"), package="TrackSOM"))
})
data <- rbindlist(data)
tracksom_res <- ConcatenateClusteringDetails(
  data,
  tracksom_result,
  timepoint.col = 'timepoint',
  timepoints = c("Mock", "SYN-1", "SYN-2", "SYN-3", "SYN-4")
)
DrawTimeseriesHeatmap(tracksom_res,
                      'timepoint',
                      c("Mock", "SYN-1", "SYN-2", "SYN-3", "SYN-4"),
                      "TrackSOM_metacluster_lineage_tracking",
                      use_cols)

}
