% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{ExportClusteringDetailsOnly}
\alias{ExportClusteringDetailsOnly}
\title{ExportClusteringDetailsOnly}
\usage{
ExportClusteringDetailsOnly(tracksom.result)
}
\arguments{
\item{tracksom.result}{TrackSOM object as generated by \code{\link{TrackSOM}}}
}
\value{
Data.table containing the meta cluster, cluster, and tracking ids for
  all the cells
}
\description{
Given a TrackSOM object export the clustering result as data.table.
}
\examples{
data_files <- sapply(c(0:4), function(i) {
system.file("extdata", paste0("synthetic_d", i, ".fcs"), package="TrackSOM")
})
use_cols <- c("x", "y", "z")
tracksom_result <- TrackSOM(inputFiles = data_files,
                            colsToUse = use_cols,
                            nClus = 10,
                            dataFileType = ".fcs"
)
tracksom_res <- ExportClusteringDetailsOnly(tracksom_result)


}
