% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionsDescramblingFunctions.R
\name{get_idx_pipe}
\alias{get_idx_pipe}
\title{Get pipe character}
\usage{
get_idx_pipe(cl)
}
\arguments{
\item{cl}{Character cluster ID}
}
\description{
Helper method to detect pipe character in meta-cluster ID
}
