% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionsDescramblingFunctions.R
\name{get_idx_roundclsbracket}
\alias{get_idx_roundclsbracket}
\title{Get round bracket}
\usage{
get_idx_roundclsbracket(cl)
}
\arguments{
\item{cl}{Character cluster ID}
}
\description{
Helper method to find round brackets in meta-cluster ID
}
