% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlotFunctions.R
\name{GetClusterOrigins}
\alias{GetClusterOrigins}
\title{GetClusterOrigins}
\usage{
GetClusterOrigins(node.dat, no_merge)
}
\arguments{
\item{node.dat}{data.table containing the nodes details}

\item{no_merge}{whether meta-cluster merging was enabled or not}
}
\value{
a vector of meta-cluster origins
}
\description{
Function to infer the origins of a meta-cluster
}
