% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlotFunctions.R
\name{GetClusterProportions}
\alias{GetClusterProportions}
\title{GetClusterProportions}
\usage{
GetClusterProportions(node.dat, dat, timepoint.col, timepoints, cluster.col)
}
\arguments{
\item{node.dat}{data.table containing the node details}

\item{dat}{the raw clustered and tracked data}

\item{timepoint.col}{the name of the column denoting the time-point}

\item{timepoints}{a vector of time-points available in the data}

\item{cluster.col}{the name of the column denoting the meta-cluster ID}
}
\value{
a vector of meta-clusters' proportion
}
\description{
Function to count the proportin of cells in each meta-cluster
}
