% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlotFunctions.R
\name{GetColourPallete}
\alias{GetColourPallete}
\title{GetColourPallete}
\usage{
GetColourPallete(standard.colours, n.col = 50, factor = NULL)
}
\arguments{
\item{standard.colours}{the colour scheme}

\item{n.col}{number of colours to get}

\item{factor}{discrete values for each colour}
}
\value{
colour mapping
}
\description{
Function to get colour pallete for network plot's nodes
}
