% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackingFunctions.R
\name{GetFirstID}
\alias{GetFirstID}
\title{GetFirstID}
\usage{
GetFirstID(mc, mcPerCell)
}
\arguments{
\item{mc}{the meta-cluster ID for each node}

\item{mcPerCell}{metaclustering per datapoint}
}
\value{
A list containing three elements:
 1) new_ids: the new IDs
 2) n: new number to be used as input for GetNewID function
 3) iter: new iteration to be used as unput to GetNewID function
}
\description{
Function which obtains the IDs for the first time point in tracking.
}
