% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlotFunctions.R
\name{GetGeomPoint}
\alias{GetGeomPoint}
\title{GetGeomPoint}
\usage{
GetGeomPoint(node.size, col.by)
}
\arguments{
\item{node.size}{whether the nodes are sized based on proportion of cells in
the meta-clusters or constant number}

\item{col.by}{what to colour the node based on}
}
\value{
a geom node point object
}
\description{
Function to get node object from
}
