% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackingFunctions.R
\name{GetNewID}
\alias{GetNewID}
\title{GetNewID}
\usage{
GetNewID(number, iteration)
}
\arguments{
\item{number}{current number of letters already issued.}

\item{iteration}{the current iteration, i.e. the number of times the letter
has been repeated (AA then iteration = 2)}
}
\value{
A list containing three elements:
  1) new_id: the new ID
  2) n: new number to be used as input for the next GetNewID function
  3) iter: new iteration to be used as input for the next GetNewID function
}
\description{
Function which gets a new ID given the number in the alphabet is next and the
iteration in the alphabet.
}
