% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionsDescramblingFunctions.R
\name{GetTransitionsAsEdges}
\alias{GetTransitionsAsEdges}
\title{GetTransitionsAsEdges}
\usage{
GetTransitionsAsEdges(dat, timepoints, timepoint_col, cluster_col)
}
\arguments{
\item{dat}{Data.table storing the clustered gated dataset}

\item{timepoints}{Vector containing the time points in order}

\item{timepoint_col}{Character Column denoting time points in dat}

\item{cluster_col}{Character Column denoting cluster id produced by tracking
by lineage determination in dat}
}
\description{
Given a TrackSOM clustering and tracking, descramble the meta-cluster ID.
Handy to work out the meta-cluster transitions.
}
\examples{

library(data.table)
data_files <- sapply(c(0:4), function(i) {
  system.file("extdata", paste0("synthetic_d", i, ".fcs"), package="TrackSOM")
})
use_cols <- c("x", "y", "z")
tracksom_result <- TrackSOM(inputFiles = data_files,
                            colsToUse = use_cols,
                            nClus = 10,
                            dataFileType = ".fcs"
)

data <- lapply(c(0:4), function(i) {
  fread(system.file("extdata", paste0("synthetic_d", i, ".csv"), package="TrackSOM"))
})
data <- rbindlist(data)
tracksom_res <- ConcatenateClusteringDetails(
  data,
  tracksom_result,
  timepoint.col = 'timepoint',
  timepoints = c("Mock", "SYN-1", "SYN-2", "SYN-3", "SYN-4")
)
GetTransitionsAsEdges(tracksom_res,
                      'timepoint',
                      c("Mock", "SYN-1", "SYN-2", "SYN-3", "SYN-4"),
                      "TrackSOM_metacluster_lineage_tracking")

}
