% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackSOM.R
\name{ParseInput}
\alias{ParseInput}
\title{ParseInput}
\usage{
ParseInput(
  dataFileType,
  inputFiles,
  compensate,
  spillover,
  transform,
  toTransform,
  transformFunction,
  scale,
  scaled.center,
  scaled.scale,
  silent,
  colsToUse
)
}
\arguments{
\item{dataFileType}{The type of input data. Can be csv, fcs, or data.frame}

\item{inputFiles}{The list containing the input data}

\item{compensate}{Whether to run data compensation}

\item{spillover}{The spillover matrix to compensate}

\item{transform}{Whether to run data transformation}

\item{toTransform}{column names or indices to transform}

\item{transformFunction}{The transform function to run}

\item{scale}{Whether to scale the data}

\item{silent}{Whether to show progress updates}

\item{scale.center}{See \link{base::scale}}

\item{scale.scale}{See \link{base::scale}}

\item{colToUse}{Which column to be used by TrackSOM}
}
\value{
FlowSOM object
}
\description{
Convert the input data to FlowSOM object ready to be processed.
Many of the parameters are used by FlowSOM's ReadInput function.
The description for these parameters are taken from FlowSOM.
}
