% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackingFunctions.R
\name{TrackingWithMerging}
\alias{TrackingWithMerging}
\title{TrackingWithMerging}
\usage{
TrackingWithMerging(fsom)
}
\arguments{
\item{fsom}{list containing a FlowSOM object and meta-clustering lists.}
}
\value{
A list containing N lists of unique IDs as determined by lineage,
  where N is no. of time steps.
}
\description{
Perform tracking of meta-clusters where merging is allowed.
Need to be run after FlowSOM's BuildSOM which contorts the SOM grid and
MetaClustering which perform meta-cluster of the SOM node.
}
\seealso{
\code{\link{TrackSOM}}.
}
