% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInput.R
\name{ParseInput}
\alias{ParseInput}
\alias{ParseInput.character}
\alias{ParseInput.list}
\alias{ParseInput.data.table}
\title{S3 ParseInput function}
\usage{
ParseInput(
  inputFiles,
  compensate,
  spillover,
  transform,
  toTransform,
  transformFunction,
  scale,
  scaled.center,
  scaled.scale,
  silent,
  colsToUse,
  ...
)

\method{ParseInput}{character}(
  inputFiles,
  compensate,
  spillover,
  transform,
  toTransform,
  transformFunction,
  scale,
  scaled.center,
  scaled.scale,
  silent,
  colsToUse,
  ...
)

\method{ParseInput}{list}(
  inputFiles,
  compensate,
  spillover,
  transform,
  toTransform,
  transformFunction,
  scale,
  scaled.center,
  scaled.scale,
  silent,
  colsToUse,
  ...
)

\method{ParseInput}{data.table}(
  inputFiles,
  compensate,
  spillover,
  transform,
  toTransform,
  transformFunction,
  scale,
  scaled.center,
  scaled.scale,
  silent,
  colsToUse,
  timepointCol,
  timepoints
)
}
\arguments{
\item{inputFiles}{The list containing the input data}

\item{compensate}{Whether to run data compensation}

\item{spillover}{The spillover matrix to compensate}

\item{transform}{Whether to run data transformation}

\item{toTransform}{column names or indices to transform}

\item{transformFunction}{The transform function to run}

\item{scale}{Whether to scale the data}

\item{silent}{Whether to show progress updates}

\item{...}{Additional arguments used if inputFiles is a data.table object}

\item{timepointCol}{A character denoting the column in data.frame that
corresponds to the time-point each cell belongs to}

\item{timepoints}{A vector of time-points.}

\item{scale.center}{See \link{base::scale}}

\item{scale.scale}{See \link{base::scale}}

\item{colToUse}{Which column to be used by TrackSOM}
}
\value{
FlowSOM object
}
\description{
Internal Function.
Convert the input data to FlowSOM object ready to be processed.
Many of the parameters are used by FlowSOM's ReadInput function.
The description for these parameters are taken from FlowSOM.
}
\section{Methods (by class)}{
\itemize{
\item \code{ParseInput(character)}: Process vector of characters input

\item \code{ParseInput(list)}: Process list of data.frames input

\item \code{ParseInput(data.table)}: Process data.table input

}}
